Patched to reduce compiler warnings (gcc-4.9)
Removed various unused variables.
Reordered various constructor member initializers.
Reduced usage of clang-specific #pragmas.

+++ core/lib/FileHandling/RINEX/RinexClockHeader.cpp
@@ -716,7 +716,7 @@ namespace gpstk
          std::list<RefClkRecord>::iterator itr = refClkList.end();
          --itr;
 
-         if ( itr->numClkRef <= itr->clocks.size() )
+         if ( itr->numClkRef <= (int)itr->clocks.size() )
          {  // Excessive # of clock references - throw
             FFStreamError e("\"ANALYSIS CLK REF\" entry exceeds "
                             "\"# of CLK REF\": " + asString(itr->numClkRef));
